+++ xporcle.js
@@ -2,12 +2,14 @@ let port = null;
 
 let onQuizPage = null;
 let interfaceBox = null;
+let interfaceContainer = null;
 let roomCode = null;
 let username = null;
 let host = null;
 let urls = {};
 let hosts = [];
 let suggestions = [];
+let voted = false;
 
 let quizStartTime = null;
 let quizRunning = false;
@@ -133,9 +135,11 @@ async function init()
 		hosts = statusResponse["hosts"];
 		suggestions = statusResponse["suggestions"];
 		saveName = statusResponse["saveName"];
+		voted = statusResponse["voted"];
 
 		const pollData =  Object.keys(statusResponse["poll_data"]).length !== 0 ? statusResponse["poll_data"] : undefined;
-		onRoomConnect(statusResponse["scores"], pollData);
+		const voteData =  Object.keys(statusResponse["vote_data"]).length !== 0 ? statusResponse["vote_data"] : undefined;
+		onRoomConnect(statusResponse["scores"], pollData, voteData);
 	}
 	else
 	{
@@ -261,7 +265,7 @@ function addInterfaceBox()
 	const gameHeader = document.querySelector(`.game-header`);
 	const staffPicks = document.querySelector(`#staff-picks-wrapper`);
 
-	let interfaceContainer = document.querySelector(`#interfaceContainer`);
+	interfaceContainer = document.querySelector(`#interfaceContainer`);
 	if (interfaceContainer === null)
 	{
 		interfaceContainer = document.createElement("div");
@@ -343,6 +347,9 @@ function resetInterface(errorElement, lastUsername, lastCode)
 	document.querySelectorAll("#startCountdown").forEach(elm => elm.remove());
 
 	document.querySelector("#pollBox")?.remove();
+	document.querySelector("#voteInfoBox")?.remove();
+	document.querySelector("#ballotPopout")?.remove();
+
 	init().then(
 		() =>
 		{
@@ -446,13 +453,20 @@ function processMessage(message)
 			{
 				document.querySelector(`#suggestQuizButton`).remove();
 				addChangeQuizButton();
-				addCreatePollButton();
 			}
 			urls = message["urls"];
 			updateHostsInLeaderboard();
 			updateLeaderboardUrls();
 			updateContextMenuHandling();
 			addSaveRoomButton();
+			if (message["poll_data"] !== null)
+			{
+				addCreatePollBox(message["poll_data"]);
+			}
+			else
+			{
+				addCreatePollButton();
+			}
 			break;
 		case "start_countdown":
 			// Start the countdown
@@ -489,6 +503,23 @@ function processMessage(message)
 		case "url_update":
 			urls[message["username"]] = message["url"]
 			updateLeaderboardUrls()
+			break;
+		case "poll_create":
+			if (document.querySelector("#pollBox") === null)
+			{
+				addCreatePollBox(message["poll_data"]);
+			}
+			break;
+		case "poll_data_update":
+			updateCreatePollBox(message["poll_data"]);
+			break;
+		case "poll_start":
+			document.querySelector("#pollBox")?.remove();
+			addVoteInfoBox(message["vote_data"]);
+			addBallotPopout(message["vote_data"].poll);
+			break;
+		case "vote_update":
+			updateVoteInfoBox(message["vote_data"]);
 			break;
 	}
 
@@ -748,7 +779,7 @@ async function loadRoom(event, form)
 
 	onRoomConnect();
 }
-function onRoomConnect(existingScores, existingPollData)
+function onRoomConnect(existingScores, existingPollData, currentVoteData)
 {
 	// Set up message handing if not done already.
 	if (!port.onMessage.hasListener(processMessage))
@@ -819,6 +850,16 @@ function onRoomConnect(existingScores, existingPollData)
 		else
 		{
 			addCreatePollButton();
+		}
+	}
+
+	// Add vote info for current poll
+	if (currentVoteData !== undefined)
+	{
+		addVoteInfoBox(currentVoteData);
+		if (!voted && !currentVoteData["finished"])
+		{
+			addBallotPopout(currentVoteData.poll);
 		}
 	}
 
@@ -941,7 +982,9 @@ function addCreatePollButton()
 	createPollButton.addEventListener("click",
 		(event) =>
 		{
-			addCreatePollBox();
+			const pollData = {duration: 30, entries: []};
+			addCreatePollBox(pollData);
+			port.postMessage({type: "poll_create", poll_data: pollData});
 			createPollButton.remove();
 		}
 	);
@@ -949,7 +992,7 @@ function addCreatePollButton()
 	interfaceBox.append(createPollButton);
 }
 
-function addCreatePollBox(pollData = {time: 30, entries: []})
+function addCreatePollBox(pollData)
 {
 	const pollBox = document.createElement("div");
 	pollBox.id = "pollBox";
@@ -1016,6 +1059,7 @@ function addCreatePollBox(pollData = {time: 30, entries: []})
 						cursor: pointer;
 						height: 1em;
 						width: 1em;
+						line:height: 1em;
 					`;
 					removeEntryButton.textContent = "×";
 					removeEntryButton.addEventListener("click",
@@ -1023,7 +1067,7 @@ function addCreatePollBox(pollData = {time: 30, entries: []})
 						{
 							newEntryListItem.remove();
 							pollData.entries = pollData.entries.filter(existingEntry => existingEntry.url !== currentQuiz.url);
-							port.postMessage({type:"pollDataUpdate", poll_data: pollData});
+							port.postMessage({type:"poll_data_update", poll_data: pollData});
 						}
 					);
 					newEntryListItem.append(removeEntryButton);
@@ -1031,7 +1075,7 @@ function addCreatePollBox(pollData = {time: 30, entries: []})
 					pollBox.querySelector("#pollEntriesList").append(newEntryListItem);
 
 					// Send updated data
-					port.postMessage({type:"pollDataUpdate", poll_data: pollData});
+					port.postMessage({type:"poll_data_update", poll_data: pollData});
 				}
 			}
 		);
@@ -1067,6 +1111,7 @@ function addCreatePollBox(pollData = {time: 30, entries: []})
 			cursor: pointer;
 			height: 1em;
 			width: 1em;
+			line-height: 1em;
 		`;
 		removeEntryButton.textContent = "×";
 		removeEntryButton.addEventListener("click",
@@ -1074,7 +1119,7 @@ function addCreatePollBox(pollData = {time: 30, entries: []})
 			{
 				entryListItem.remove();
 				pollData.entries = pollData.entries.filter(existingEntry => existingEntry.url !== entry.url);
-				port.postMessage({type:"pollDataUpdate", poll_data: pollData});
+				port.postMessage({type:"poll_data_update", poll_data: pollData});
 			}
 		);
 		entryListItem.append(removeEntryButton);
@@ -1082,32 +1127,33 @@ function addCreatePollBox(pollData = {time: 30, entries: []})
 	}
 	pollBox.append(pollEntriesList);
 
-	const pollTimeInput = document.createElement("input");
-	pollTimeInput.type = "number";
-	pollTimeInput.min = "10";
-	pollTimeInput.max = "60";
-	pollTimeInput.value = pollData.time;
-	pollTimeInput.addEventListener("change",
+	const pollDurationInput = document.createElement("input");
+	pollDurationInput.id = "pollDuration";
+	pollDurationInput.type = "number";
+	pollDurationInput.min = "10";
+	pollDurationInput.max = "60";
+	pollDurationInput.value = pollData.duration;
+	pollDurationInput.addEventListener("change",
 		(event) =>
 		{
-			pollData.time = Math.max(10, Math.min(pollTimeInput.value, 60));
+			pollData.duration = Math.max(10, Math.min(pollDurationInput.value, 60));
 			// Send updated data
-			port.postMessage({type:"pollDataUpdate", poll_data: pollData});
+			port.postMessage({type:"poll_data_update", poll_data: pollData});
 		}
 	);
-	pollTimeInput.addEventListener("blur",
+	pollDurationInput.addEventListener("blur",
 		(event) =>
 		{
-			pollTimeInput.value = pollData.time;
+			pollDurationInput.value = pollData.duration;
 		}
 	);
-	const pollTimeContainer = document.createElement("div");
-	pollTimeContainer.append(
+	const pollDurationContainer = document.createElement("div");
+	pollDurationContainer.append(
 		document.createTextNode("Poll Time:"),
-		pollTimeInput,
+		pollDurationInput,
 		document.createTextNode("s")
 	);
-	pollBox.append(pollTimeContainer);
+	pollBox.append(pollDurationContainer);
 
 	if (pollData.entries.length > 1)
 	{
@@ -1120,8 +1166,7 @@ function addCreatePollBox(pollData = {time: 30, entries: []})
 				// Send poll to server
 				port.postMessage(
 					{
-						type:"startPoll",
-						poll_data: pollData,
+						type:"poll_start",
 						start_time: (new Date()).getTime()
 					}
 				);
@@ -1131,7 +1176,312 @@ function addCreatePollBox(pollData = {time: 30, entries: []})
 		pollBox.append(sendPollToRoomButton);
 	}
 
-	interfaceContainer.append(pollBox);
+	interfaceContainer.insertBefore(pollBox, interfaceBox.nextElementSibling);
+}
+
+function updateCreatePollBox(pollData)
+{
+	const pollBox = document.querySelector("#pollBox");
+	// Clear entires
+	pollBox.querySelectorAll("ol li").forEach(entry => entry.remove());
+	// Add updated entries
+	for (const entry of pollData.entries)
+	{
+		const entryListItem = document.createElement("li");
+		entryListItem.style =
+		`
+			display: grid;
+			grid-template-columns: 1fr 1em;
+		    align-items: center;
+		`;
+		entryListItem.id = entry.short_title;
+		entryListItem.textContent = entry.short_title;
+		removeEntryButton = document.createElement("div");
+		removeEntryButton.style =
+		`
+			background-color: red;
+			color: white;
+			border-radius: 50%;
+			display: flex;
+			justify-content: center;
+			align-content: center;
+			cursor: pointer;
+			height: 1em;
+			width: 1em;
+			line-height: 1em;
+		`;
+		removeEntryButton.textContent = "×";
+		removeEntryButton.addEventListener("click",
+			(event) =>
+			{
+				entryListItem.remove();
+				pollData.entries = pollData.entries.filter(existingEntry => existingEntry.url !== entry.url);
+				port.postMessage({type:"poll_data_update", poll_data: pollData});
+			}
+		);
+		entryListItem.append(removeEntryButton);
+		pollBox.querySelector("ol").append(entryListItem);
+	}
+
+	// Update duration
+	pollBox.querySelector("#pollDuration").value = pollData.duration;
+
+	// Update start button display
+	let sendPollToRoomButton = document.querySelector("#sendPollToRoom");
+	if (pollData.entries.length > 1)
+	{
+		if (sendPollToRoomButton === null)
+		{
+			sendPollToRoomButton = document.createElement("Button");
+			sendPollToRoomButton.textContent = "Send Poll to Room";
+			sendPollToRoomButton.id = "sendPollToRoom";
+			sendPollToRoomButton.addEventListener("click",
+				(event) =>
+				{
+					// Send poll to server
+					port.postMessage(
+						{
+							type:"poll_start",
+							poll_data: pollData,
+							start_time: (new Date()).getTime()
+						}
+					);
+					pollBox.remove();
+				}
+			);
+			pollBox.append(sendPollToRoomButton);
+		}
+	}
+	else
+	{
+		sendPollToRoomButton?.remove();
+	}
+
+}
+
+function addVoteInfoBox(voteData)
+{
+	document.querySelector("#voteInfoBox")?.remove()
+	const voteInfoBox = document.createElement("div");
+	voteInfoBox.id = "voteInfoBox";
+	voteInfoBox.style = 
+	`
+		width: calc(((100vw - 960px) / 2) - 10px);
+		padding: 0.5em;
+		box-sizing: border-box;
+		max-width: 400px;
+		list-style: none;
+		border-width: 1px;
+		border-style: solid;
+		border-color: darkgrey;
+		border-radius: 0.25em;
+		background-color: white;
+		display: grid;
+		row-gap: 1em;
+		grid-template-columns: 100%;
+	`;
+
+	const header = document.createElement("h2");
+	header.textContent = "Vote Status";
+	voteInfoBox.append(header);
+
+	const timer = document.createElement("p");
+	timer.append(document.createTextNode("Time left: "));
+	const timeLeft = document.createElement("span");
+	timeLeft.id = "timeLeft";
+	if (!voteData["finished"])
+	{
+		timeLeft.textContent = Math.floor((voteData.start_time + voteData.duration*1000 - (new Date()).getTime())/1000);
+		const intervalId = setInterval(
+			() =>
+			{
+				if (timeLeft.textContent !== "Finished")
+				{
+					timeLeft.textContent = ((voteData.start_time + voteData.duration*1000 - (new Date()).getTime())/1000).toFixed(1);
+				}
+				else
+				{
+					clearInterval(intervalId);
+				}
+			},
+			100
+		);
+		setTimeout(
+			() =>
+			{
+				clearInterval(intervalId);
+				timeLeft.textContent = "Finished";
+			},
+			voteData.duration*1000
+		);
+	}
+	else
+	{
+		timeLeft.textContent = "Finished";
+	}
+	timer.append(timeLeft);
+	voteInfoBox.append(timer);
+
+	const responses = document.createElement("p");
+	const responseCount = document.createElement("span");
+	responseCount.id = "responseCount";
+	responses.append(responseCount);
+	responses.append(document.createTextNode(" responded"));
+	voteInfoBox.append(responses);
+
+	if (!voteData.finished)
+	{
+		const openBallotButton = document.createElement("button");
+		openBallotButton.id = "openBallotButton";
+		openBallotButton.textContent = "Open Vote Box";
+		openBallotButton.addEventListener("click", () => {addBallotPopout(voteData.poll)})
+		voteInfoBox.append(openBallotButton);
+	}
+
+	interfaceContainer.append(voteInfoBox);
+
+	updateVoteInfoBox(voteData);
+}
+
+function updateVoteInfoBox(voteData)
+{
+	let voteInfoBox = document.querySelector("#voteInfoBox");
+	if (voteInfoBox === null)
+	{
+		return addVoteInfoBox(voteData);
+	}
+
+	document.querySelector("#responseCount").textContent = `${voteData.response_count}/${voteData.num_voters}`;
+	if (voteData.finished)
+	{
+		const winnerHeader = document.createElement("h3");
+		winnerHeader.textContent = "Winner:";
+		voteInfoBox.append(winnerHeader);
+		const winnerLink = document.createElement("a");
+		winnerLink.href = voteData.winner.url;
+		winnerLink.textContent = voteData.winner.short_title;
+		voteInfoBox.append(winnerLink);
+
+		voteInfoBox.querySelector("#timeLeft").textContent = "Finished";
+
+		voteInfoBox.querySelector("#openBallotButton")?.remove();
+	}
+}
+
+function addBallotPopout(poll)
+{
+	document.querySelector("#ballotPopout")?.remove();
+	const ballotPopout = document.createElement("div");
+	ballotPopout.id = "ballotPopout";
+	ballotPopout.style =
+	`
+		position: fixed;
+		background: white;
+		left: 50%;
+		top: 50%;
+		transform: translate(-50%,-50%);
+		width: 40vw;
+		min-height: 40vw;
+		z-index: 999999;
+		border-radius: 0.5em;
+		border: 1px solid #888;
+		display: grid;
+		row-gap: 1em;
+		grid-template-columns: 100%;
+		padding: 0.5em;
+	`;
+
+	const closeBoxButton = document.createElement("button");
+	closeBoxButton.textContent = "×";
+	closeBoxButton.style =
+	`
+		appearance: none;
+		padding: 0;
+		border: 0;
+		position: absolute;
+		right: 1rem;
+		top: 1rem;
+		color: grey;
+		cursor: pointer;
+		background: none;
+		font-size: 150%;
+	`;
+	closeBoxButton.addEventListener("click",
+		(event) =>
+		{
+			ballotPopout.remove();
+		}
+	);
+	ballotPopout.append(closeBoxButton);
+
+	const header = document.createElement("h1");
+	header.textContent = "Next Quiz Poll";
+	ballotPopout.append(header);
+
+	const timer = document.createElement("p");
+	timer.append(document.createTextNode("Time left: "));
+	const timeLeft = document.createElement("span");
+	timeLeft.id = "timeLeft";
+	timeLeft.textContent = Math.floor((poll.start_time + poll.duration*1000 - (new Date()).getTime())/1000);
+	const intervalId = setInterval(
+		() =>
+		{
+			timeLeft.textContent = ((poll.start_time + poll.duration*1000 - (new Date()).getTime())/1000).toFixed(1);
+		},
+		100
+	);
+	setTimeout(
+		() =>
+		{
+			clearInterval(intervalId);
+			timeLeft.textContent = "Finished";
+			// disable all checkboxes and submit button
+			Array.from(ballotPopout.querySelectorAll(`input[type="checkbox"], #submitBallot`)).forEach(elem => elem.disabled = true);
+			setTimeout(()=>ballotPopout.remove(), 2000);
+		},
+		poll.duration*1000
+	);
+	timer.append(timeLeft);
+	ballotPopout.append(timer);
+
+	const entriesList = document.createElement("ol");
+	entriesList.style = 
+	`
+		padding: 0;
+	`;
+	for (const entry of poll.entries)
+	{
+		const entryListItem = document.createElement("li");
+		entryListItem.textContent = entry.long_title;
+		entryListItem.style =
+		`
+			display: grid;
+			grid-template-columns: 1fr 1em;
+		    align-items: center;
+		`;
+
+		const checkbox = document.createElement("input");
+		checkbox.type = "checkbox";
+		entryListItem.append(checkbox)
+		entriesList.append(entryListItem);
+	}
+	ballotPopout.append(entriesList);
+
+	const submitButton = document.createElement("button");
+	submitButton.textContent = "Submit";
+	submitButton.id = "submitBallot";
+	submitButton.addEventListener("click",
+		(event) =>
+		{
+			const votes = Array.from(ballotPopout.querySelectorAll(`input[type="checkbox"]`)).map(checkbox => Number(checkbox.checked));
+			port.postMessage({type: "poll_vote", votes: votes});
+			ballotPopout.remove();
+			document.querySelector("#openBallotButton")?.remove();
+		}
+	);
+	ballotPopout.append(submitButton);
+
+	document.body.insertBefore(ballotPopout, document.body.firstElementChild);
 }
 
 async function saveRoom(event)
@@ -1877,7 +2227,7 @@ function displayContextMenu(event)
 	menuItem.addEventListener("click",
 		(event) =>
 		{
-			port.postMessage({type: "host_promotion", username: targetUsername});
+			port.postMessage({type: "make_host", username: targetUsername});
 			removeContextMenu();
 		}
 	);