+++ CHANGELOG.md
@@ -3,9 +3,11 @@ Changelog
 
 [Unreleased]
 ------------
-###Added
+### Added
 - Poll feature to vote for next quiz.
-  - Interface for hosts to create a poll together
+  - Interface for hosts to create a poll together.
+  - Popup box for users to submit their votes.
+  - Vote progress information box.
 
 [v2.0.0] - 2021-03-22
 ---------------------